Random forest classifier training and prediction
Changed training area to be larger. Removed conversion of transition_map to FeatureCollection as it is not neede for Random Forest. Added a band to train_img that corresponds to the transition_map. Sample points from train_img. Train random forest classifier, then predict test image and plot. Training is extremely fast and predicted transition map is 100% deforested (bug).